pyproject.toml
[project]
name = "django-drf"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.3",
    "django-filter>=25.2",
    "djangorestframework>=3.17.1",
]
